requirements.txt
bcrypt==4.0.1
greenlet==2.0.2
PyJWT==2.8.0
PyMySQL==1.1.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
